feat: add api endpoint tests
